Refac & CleanUp: Schema
