+++ src/schema/resolvers/queries.js
@@ -1,0 +1,152 @@
+const path = require('path')
+const fs = require('fs')
+
+const git = require('isomorphic-git')
+git.plugins.set('fs', fs)
+
+const folders = require('../../functions/folder')
+const files = require('../../functions/file')
+
+const getFolderSize = require('../../utils/getFolderSize')
+const { getRelFilePath, repoDir } = require('../../utils/parsePath')
+const { checkoutBranch } = require('../../functions/git.js')
+
+const { PubSub } = require('apollo-server')
+const pubsub = new PubSub()
+
+const FILE_OPENED = 'FILE_OPENED'
+
+const resolvers = {
+	Subscription: {
+		openFileSub: {
+			subscribe: () => pubsub.asyncIterator([FILE_OPENED]),
+		},
+	},
+	Query: {
+		getNestedFolders: async (_, args) => {
+			if (fs.existsSync(args.path)) {
+				const data = await folders
+					.getNestedFolders(args.path)
+					.then(response => response)
+				const withParent = {
+					name: path.parse(args.path).name,
+					path: args.path,
+					children: data,
+				}
+				return withParent
+			}
+			return new Error('ENOENT')
+		},
+		getFolderWithFiles: async (_, args) => {
+			if (fs.existsSync(args.path)) {
+				const data = await folders
+					.getFolderWithFiles(args.path)
+					.then(response => response)
+				const folderSize = await getFolderSize(args.path)
+					.filter(Boolean)
+					.map(file => fs.readFileSync(file))
+					.join('\n')
+				const withParent = {
+					name: path.parse(args.path).name,
+					type: 'folder',
+					path: args.path,
+					size: folderSize.length,
+					children: data,
+					createdAt: fs.statSync(args.path).birthtime,
+				}
+				return withParent
+			}
+			return new Error('ENOENT')
+		},
+		getFile: async (_, args) => {
+			if (fs.existsSync(args.path)) {
+				return files
+					.getFile(args.path)
+					.then(success => success)
+					.catch(failure => new Error(failure))
+			}
+			return new Error('ENOENT')
+		},
+		openFile: (_, args) => {
+			if (fs.existsSync(args.path)) {
+				return files
+					.getFile(args.path)
+					.then(success => {
+						pubsub.publish(FILE_OPENED, { openFileSub: success })
+						return success
+					})
+					.catch(failure => new Error(failure))
+			}
+			return new Error('ENOENT')
+		},
+		searchFiles: (_, args) =>
+			files
+				.searchFiles(args.fileName)
+				.then(data => data)
+				.catch(e => e),
+		getCommitLog: (_, { path: repoDir }) => {
+			return git
+				.log({
+					dir: repoDir,
+					depth: 10,
+					ref: 'master',
+				})
+				.then(list => list)
+				.catch(error => new Error(error))
+		},
+		getCommit: (_, { id, path: repoDir }) => {
+			return git
+				.readObject({
+					dir: repoDir,
+					oid: id,
+				})
+				.then(({ object }) => object)
+				.catch(error => new Error(error))
+		},
+		getCommits: async (_, { path, commits }) => {
+			const results = await commits.map(commit =>
+				git
+					.readObject({
+						dir: path,
+						oid: commit,
+					})
+					.then(({ object }) => object)
+					.catch(error => new Error(error))
+			)
+			return Promise.all(results).then(data => data)
+		},
+		getCommitContent: (_, { id, path }) => {
+			return git
+				.readObject({
+					dir: repoDir(path),
+					oid: id,
+					filepath: getRelFilePath(path),
+					encoding: 'utf8',
+				})
+				.then(({ object }) => object)
+				.catch(error => new Error(error))
+		},
+		showFilesInBranch: (_, { branchName, appName, entity }) => {
+			return checkoutBranch(
+				branchName,
+				'./../apps/' + appName + '/data/' + entity
+			)
+				.then(() => {
+					return resolvers.Query.getFolderWithFiles('', {
+						path: './../apps/' + appName + '/data/' + entity,
+					}).then(data => {
+						return data
+					})
+				})
+				.then(data => {
+					checkoutBranch(
+						'master',
+						'./../apps/' + appName + '/data/' + entity
+					)
+					return data
+				})
+		},
+	},
+}
+
+module.exports = resolvers
