+++ src/schema/types/index.js
@@ -1,0 +1,11 @@
+const queryTypes = require('./queries/index')
+const mutationTypes = require('./mutations/index')
+const subscriptionTypes = require('./subscriptions/index')
+const scalarTypes = require('./scalars/index')
+
+module.exports = {
+	queryTypes,
+	mutationTypes,
+	subscriptionTypes,
+	scalarTypes,
+}
