+++ src/schema/types/subscriptions/index.js
@@ -1,0 +1,9 @@
+const { gql } = require('apollo-server-express')
+
+const typeDefs = gql`
+	type Subscription {
+		openFileSub: File
+	}
+`
+
+module.exports = typeDefs
